+++ src/main/java/com/example/sns/auth/application/AuthService.java
@@ -2,7 +2,6 @@ package com.example.sns.auth.application;
 
 import com.example.sns.auth.application.dto.OAuthUserInfoDto;
 import com.example.sns.auth.infrastructure.GoogleClient;
-import com.example.sns.auth.infrastructure.JwtProvider;
 import com.example.sns.auth.presentation.dto.TokenResponse;
 import com.example.sns.member.domain.Member;
 import com.example.sns.member.domain.MemberRepository;
