+++ src/main/java/com/example/sns/auth/application/JwtProvider.java
@@ -1,4 +1,4 @@
-package com.example.sns.auth.infrastructure;
+package com.example.sns.auth.application;
 
 import com.example.sns.auth.exception.ExpiredTokenException;
 import com.example.sns.auth.exception.InvalidTokenException;
