+++ src/test/java/com/example/sns/auth/application/JwtProviderTest.java
@@ -1,5 +1,6 @@
-package com.example.sns.auth.infrastructure;
+package com.example.sns.auth.application;
 
+import com.example.sns.auth.application.JwtProvider;
 import com.example.sns.auth.exception.ExpiredTokenException;
 import com.example.sns.auth.exception.InvalidTokenException;
 import org.junit.jupiter.api.BeforeEach;
